backend/requirements.txt
# FastAPI Backend Requirements (Python 3.13 compatible)

# Web Framework
fastapi>=0.128.0
uvicorn[standard]>=0.40.0

# Database
sqlalchemy>=2.0.46
greenlet>=3.2.4
alembic>=1.18.0
asyncpg>=0.31.0
psycopg2-binary>=2.9.11

# Redis
redis>=7.1.0

# Authentication
pyjwt>=2.11.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=25.1.0

# Validation & Serialization
pydantic>=2.12.0
pydantic-settings>=2.12.0
email-validator>=2.3.0

# WebSocket
python-socketio>=5.16.0

# HTTP Client (for LINE API)
httpx>=0.28.0

# Email
aiosmtplib>=4.0.0

# Utilities
python-multipart>=0.0.22
python-dotenv>=1.2.0
structlog>=25.0.0

# Testing
pytest>=9.0.0
pytest-asyncio>=1.3.0
